move createPeerDescriptor
